+++ src/Common/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The change log describes what is "Added", "Removed", "Changed" or "Fixed" between each release.
 
+## 2.1
+
++ Add to coordinates altitude
++ Add measuring distance between two coordinates
++ Fix potential issue in findPlaceByCoordinates method
++ Implement functionality for find neighbor places
+
 ## 2.0
 
 + Transform to common part of library